update transfer function calculation
Previously, the transfer function calculation VI was a bit weird and the
transfer function was not just y=mx+c. We've made it into this (as you
can see in "Apply Transfer Function.vi". Also, we added a "dead zone"
when passing in 0 to this VI, so passing in 0 can never cause the robot
to move slightly. This is necessary because the intercept is never quite
correct.

Wall following works really well here, so if you want to showcase proper
wall following, revert here.